+++ src/liblightmetrica/pathtrace.cpp
@@ -102,7 +102,7 @@ bool PathtraceRenderer::Impl::Configure( const ConfigNode& node, const Assets& a
 bool PathtraceRenderer::Impl::Render( const Scene& scene )
 {
 	auto* masterFilm = scene.MainCamera()->GetFilm();
-	std::atomic<int> processedBlocks(0);
+	std::atomic<long long> processedBlocks(0);
 
 	signal_ReportProgress(0, false);
